pyproject.toml
[project]
name = "synthetic-data"
version = "0.1.0"
description = "Synthetic data generator with OpenAI and evaluation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "openai>=2.6.0",
    "pandas>=2.3.3",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
]